eval/requirements.txt
torch
transformers
pyyaml 
git+https://github.com/felipemaiapolo/tinyBenchmarks # tinyMMLU